Move Pruning to offline_transformations target
