+++ inference-engine/tests/functional/inference_engine/CMakeLists.txt
@@ -17,6 +17,7 @@ set(LINK_LIBRARIES
     openvino::conditional_compilation
     sharedTestClasses
     inference_engine_snippets
+    offline_transformations
 )
 
 set(DEPENDENCIES
